Encapsulated spotify wrappers in spotify.js
